Login

+++ models/user.go
@@ -1,9 +1,9 @@
 package models
 
 type User struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
-	Role     int    `json:"role"`
+	Role     int64  `json:"role"`
 }
